- Moved code - Even more renaming - Implemented "drawTextIllsmouth"
git-svn-id: https://svn.scummvm.org:4444/svn/comet@133 e395a8f9-5cf0-422e-971d-9a083d8f24e8

+++ engines/comet/scene.h
@@ -63,9 +63,9 @@ public:
 	void addBlockingRect(int x1, int y1, int x2, int y2);
 	void removeBlockingRect(int x, int y);
 
-	int checkCollisionWithBounds(const Common::Rect &collisionRect, int direction);
-	int checkCollisionWithExits(const Common::Rect &collisionRect, int direction);
-	int checkCollisionWithBlockingRects(Common::Rect &collisionRect, Common::Rect &obstacleRect);
+	uint16 checkCollisionWithBounds(const Common::Rect &collisionRect, int direction);
+	uint16 checkCollisionWithExits(const Common::Rect &collisionRect, int direction);
+	uint16 checkCollisionWithBlockingRects(Common::Rect &collisionRect, Common::Rect &obstacleRect);
 
 	void getExitRect(int index, int &x1, int &y1, int &x2, int &y2);
 	void findExitRect(int sceneNumber, int moduleNumber, int direction, int &x1, int &y1, int &x2, int &y2, int &outDirection);
